organize hash.go
