+++ crypto/merkle_test.go
@@ -1,5 +1,8 @@
 package crypto
 
+// TODO: Give this file a lot more love. And maybe break it into its own
+// package.
+
 import (
 	"bytes"
 	"crypto/rand"
